System.Text.Json support

+++ src/Disqord.Voice.Api/Models/Payloads/HelloJsonModel.cs
@@ -5,5 +5,5 @@ namespace Disqord.Voice.Api.Models;
 public class HelloJsonModel : JsonModel
 {
     [JsonProperty("heartbeat_interval")]
-    public int HeartbeatInterval;
+    public double HeartbeatInterval;
 }
